refactor: use analytics hook in nextjs full-stack sample
