+++ examples/nextjs/full-stack/sample-app/src/pages/index.tsx
@@ -1,6 +1,6 @@
 import Head from 'next/head';
 import { useState } from 'react';
-import { getBrowserAnalytics } from '@/lib/rudderstack-client';
+import useRudderAnalytics from '@/useRudderAnalytics';
 import {
   destinationRouting,
   eventTypes,
@@ -14,11 +14,11 @@ type HomeProps = {
 };
 
 export default function Home({ serverPageEvent }: HomeProps) {
+  const analytics = useRudderAnalytics();
   const [browserStatus, setBrowserStatus] = useState('No browser event sent yet.');
   const [serverStatus, setServerStatus] = useState(serverPageEvent);
 
   const sendBrowserEvent = (type: EventType) => {
-    const analytics = getBrowserAnalytics();
     if (!analytics) {
       setBrowserStatus('Add the browser environment variables before you send events.');
       return;
